fix(astro): fixing issue with not reporting successful process
There is an issue regarding the e2e tests using cypress and the `dev` target that cannot be able to start propertly.

The executor `@nrwl/cypress` uses the `devServerTarget` to run the target first and then boot cypress. You can see it here https://github.com/nrwl/nx/blob/master/packages/cypress/src/executors/cypress/cypress.impl.ts#L143

There was no successful response so cypress was unable to start it correctly.

Reviewing the code, I found that the regex expression to resolve the promise was not quite right. Instead of using "started" which is no longer the output of Astro, I change the value to "Server started".




Due to changes in the regex on the  the `dev` executor the ``was unable to detect that the `devServerTarget`